+++ tests/handling/test_cause_logging.py
@@ -5,7 +5,7 @@ import pytest
 
 import kopf
 from kopf.reactor.causation import ALL_REASONS, HANDLER_REASONS, Reason
-from kopf.reactor.handling import process_resource_event
+from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
