+++ tests/handling/test_multistep.py
@@ -4,7 +4,7 @@ import pytest
 
 import kopf
 from kopf.reactor.causation import Reason, HANDLER_REASONS
-from kopf.reactor.handling import process_resource_event
+from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
