+++ tests/handling/test_timing_consistency.py
@@ -4,7 +4,7 @@ import datetime
 import freezegun
 
 import kopf
-from kopf.reactor.handling import process_resource_event
+from kopf.reactor.processing import process_resource_event
 from kopf.structs.containers import ResourceMemories
